Edit view livewire crm product

+++ resources/views/livewire/crm/product/product-lists.blade.php
@@ -23,14 +23,23 @@
                     </div>
                     <div class="col-3 d-flex justify-content-center">
                         <div class="btn-group w-100" role="group">
-                            <button wire:click="$dispatch('refresh-product-ax')" type="button"
-                                class="btn btn-secondary" data-toggle="modal" data-target="#modal-product-ax">
-                                <i class="fas fa-plus"></i> AX
-                            </button>
-                            <button type="button" class="btn btn-info" data-toggle="modal"
-                                data-target="#modal-add-product">
-                                <i class="fas fa-plus"></i> Add
-                            </button>
+
+                            @can('product.create')
+                                <button wire:click="$dispatch('refresh-product-ax')" type="button"
+                                    class="btn btn-secondary" data-toggle="modal" data-target="#modal-product-ax">
+                                    <i class="fas fa-plus"></i> AX
+                                </button>
+
+                                {{-- @livewire('crm.product.product-ax-lists') --}}
+
+                                <button type="button" class="btn btn-info" data-toggle="modal"
+                                    data-target="#modal-add-product">
+                                    <i class="fas fa-plus"></i> Add
+                                </button>
+
+                                {{-- @livewire('crm.product.product-create') --}}
+                            @endcan
+
                             <button wire:click="$dispatch('refresh-product')" type="button" class="btn btn-success">
                                 <i class="fas fa-sync-alt"></i>
                             </button>
@@ -54,8 +63,8 @@
                 </div>
 
                 <div class="table-responsive">
-                    <table class="table table-hover table-sm">
-                        <thead>
+                    <table class="table table-sm table-hover table-bordered">
+                        <thead class="table-info">
                             <th scope="col">#</th>
                             <th scope="col">Created</th>
                             <th scope="col">Updated</th>
@@ -64,9 +73,15 @@
                             <th scope="col">Code</th>
                             <th scope="col">Name</th>
                             <th scope="col">Brand</th>
-                            <th scope="col">Supplier Rep.</th>
+                            <th scope="col">Supplier</th>
                             <th scope="col">Principal</th>
-                            <th scope="col" style="width: 115px">Action</th>
+                            @can('crm.create')
+                                <th scope="col" colspan="3" class="text-center">Action</th>
+                            @elsecan('product.edit')
+                                <th scope="col" colspan="3" class="text-center">Action</th>
+                            @elsecan('product.delete')
+                                <th scope="col" colspan="3" class="text-center">Action</th>
+                            @endcan
                         </thead>
 
                         <tbody>
@@ -109,7 +124,7 @@
                                         </div>
 
                                     </td>
-                                    <td>
+                                    <td class="text-center">
                                         @if ($item->source === '0')
                                             <span class="badge badge-pill badge-primary">
                                                 AX
@@ -130,35 +145,68 @@
                                     <td>{{ $item->brand }}</td>
                                     <td>{{ $item->supplier_rep }}</td>
                                     <td>{{ $item->principal }}</td>
-                                    <td>
-                                        <button
-                                            wire:click.prevent="$dispatch('select-product',{id:{{ $item->id }}})"
-                                            wire:click="$dispatch('refresh-product')" class="btn btn-success btn-sm">
-                                            <i class="fas fa-check"></i>
-                                        </button>
-                                        @if ($item->source === '1' || $item->source === '2')
-                                            <button
-                                                wire:click.prevent="$dispatch('edit-product',{id:{{ $item->id }}})"
-                                                type="button" class="btn btn-secondary btn-sm" data-toggle="modal"
-                                                data-target="#modal-edit-product">
-                                                <i class="fas fa-edit"></i>
-                                            </button>
-                                            <button
-                                                wire:click.prevent="deleteProduct({{ $item->id }},{{ "'" . str_replace("'", '', $item->product_name) . "'" }})"
-                                                {{-- wire:click.prevent="deleteProduct({{ $item->id }},{{ "'" . $item->product_name . "'" }})" --}} class="btn btn-sm btn-danger">
-                                                <i class="fas fa-trash"></i>
-                                            </button>
-                                        @endif
-                                    </td>
+
+                                    @if ($item->source === '1' || $item->source === '2')
+                                        @can('crm.create')
+                                            <td style="width: 40px" class="p-1 text-center">
+                                                <button
+                                                    wire:click.prevent="$dispatch('select-product',{id:{{ $item->id }}})"
+                                                    class="btn btn-success btn-sm">
+                                                    <i class="fas fa-check"></i>
+                                                </button>
+                                            </td>
+                                        @endcan
+
+                                        @can('product.edit')
+                                            <td style="width: 40px" class="p-1 text-center">
+                                                <button
+                                                    wire:click.prevent="$dispatch('edit-product',{id:{{ $item->id }}})"
+                                                    type="button" class="btn btn-secondary btn-sm" data-toggle="modal"
+                                                    data-target="#modal-edit-product">
+                                                    <i class="fas fa-edit"></i>
+                                                </button>
+                                            </td>
+                                        @endcan
+
+                                        @can('product.delete')
+                                            <td style="width: 40px" class="p-1 text-center">
+                                                <button
+                                                    wire:click.prevent="deleteProduct({{ $item->id }},{{ "'" . str_replace("'", '', $item->product_name) . "'" }})"
+                                                    class="btn btn-sm btn-danger">
+                                                    <i class="fas fa-trash"></i>
+                                                </button>
+                                            </td>
+                                        @endcan
+                                    @else
+                                        {{-- Data Source = 0 AX --}}
+                                        @can('crm.create')
+                                            <td colspan="3" style="width: 40px" class="p-1 text-center">
+                                                <button
+                                                    wire:click.prevent="$dispatch('select-product',
+                                            {id:{{ $item->id }}})"
+                                                    wire:click="$dispatch('refresh-product')"
+                                                    class="btn btn-success btn-sm">
+                                                    <i class="fas fa-check"></i>
+                                                </button>
+                                            </td>
+                                        @elsecan('product.edit')
+                                            <td colspan="3"></td>
+                                        @elsecan('product.delete')
+                                            <td></td>
+                                        @endcan
+                                    @endif
                                 </tr>
                             @endforeach
 
                             @livewire('crm.product.product-edit')
 
-                            {{-- @livewire('crm.product-edit') --}}
-
                         </tbody>
                     </table>
+
+                    @livewire('crm.product.product-ax-lists')
+
+                    @livewire('crm.product.product-create')
+
                 </div>
 
             </div>
